+++ src/main/java/com/hankcs/hanlp/mining/word2vec/Utility.java
@@ -1,12 +1,13 @@
 
 package com.hankcs.hanlp.mining.word2vec;
 
+import java.io.*;
+
 /**
  * 一些工具方法
  */
 final class Utility
 {
-
     private static final int SECOND = 1000;
     private static final int MINUTE = 60 * SECOND;
     private static final int HOUR = 60 * MINUTE;
@@ -43,5 +44,77 @@ final class Utility
 //        text.append(ms + " ms");
 
         return text.toString();
+    }
+
+    /**
+     * @param c
+     */
+    public static void closeQuietly(Closeable c)
+    {
+        try
+        {
+            if (c != null) c.close();
+        }
+        catch (IOException ignored)
+        {
+        }
+    }
+
+    /**
+     * @param raf
+     */
+    public static void closeQuietly(RandomAccessFile raf)
+    {
+        try
+        {
+            if (raf != null) raf.close();
+        }
+        catch (IOException ignored)
+        {
+        }
+    }
+
+    public static void closeQuietly(InputStream is)
+    {
+        try
+        {
+            if (is != null) is.close();
+        }
+        catch (IOException ignored)
+        {
+        }
+    }
+
+    public static void closeQuietly(Reader r)
+    {
+        try
+        {
+            if (r != null) r.close();
+        }
+        catch (IOException ignored)
+        {
+        }
+    }
+
+    public static void closeQuietly(OutputStream os)
+    {
+        try
+        {
+            if (os != null) os.close();
+        }
+        catch (IOException ignored)
+        {
+        }
+    }
+
+    public static void closeQuietly(Writer w)
+    {
+        try
+        {
+            if (w != null) w.close();
+        }
+        catch (IOException ignored)
+        {
+        }
     }
 }
