+++ src/main/java/com/hankcs/hanlp/mining/word2vec/Word2VecTraining.java
@@ -110,7 +110,7 @@ class Word2VecTraining
                             System.err.printf("%cAlpha: %f  iter: %d  Progress: %.2f%%  Words/thread/sec: %.2fk", 13, alpha, local_iter,
                                               percent * 100,
                                               wordCountActual / (float) (cost_time));
-                            String etd = Utils.humanTime((long) (cost_time / percent * (1.f - percent)));
+                            String etd = Utility.humanTime((long) (cost_time / percent * (1.f - percent)));
                             if (etd.length() > 0) System.err.printf("  ETD: %s", etd);
                             System.err.flush();
                         }
@@ -360,7 +360,7 @@ class Word2VecTraining
         }
 
         System.err.println();
-        logger.info(String.format("finished training in %s", Utils.humanTime(System.currentTimeMillis() - timeStart)));
+        logger.info(String.format("finished training in %s", Utility.humanTime(System.currentTimeMillis() - timeStart)));
         // lose weight
         syn1 = null;
         table = null;
@@ -391,9 +391,9 @@ class Word2VecTraining
         finally
         {
             corpus.close();
-            Utils.closeQuietly(pw);
-            Utils.closeQuietly(w);
-            Utils.closeQuietly(os);
+            Utility.closeQuietly(pw);
+            Utility.closeQuietly(w);
+            Utility.closeQuietly(os);
         }
     }
